+++ backend/views/plan/MyPlans.ejs
@@ -37,10 +37,8 @@
   <script>
       function openPop(i) {
         var title = document.getElementsByClassName(String(i))[0].innerHTML;
-        alert(title);
         var pop_title = "스케줄표";
         window.open("/myplans/schedule?title="+title, pop_title, "width=1000, height=800");
-
       }
   </script>
 </html>
